+++ app.py
@@ -1,6 +1,6 @@
 import os
 
-from app import create_app
+from flask_app import create_app
 
-config_name = os.getenv('CFG') or 'dev'
+config_name = os.getenv('APP_CFG') or 'dev'
 app = create_app(config_name)
