--- app/config/__init__.py
@@ -1,4 +1,0 @@
-from .dev import DevConfig
-from .dev import TestConfig
-
-app_config = {'dev': DevConfig, 'test': TestConfig}
