+++ flask_app/config/__init__.py
@@ -1,0 +1,5 @@
+from .dev import DevConfig
+from .dev import TestConfig
+
+app_config = {'dev': DevConfig,
+              'test': TestConfig}
