+++ flask_app/models/user.py
@@ -3,7 +3,7 @@ import os
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
 
-from app.extensions import db
+from flask_app.extensions import db
 
 
 class User(db.Model, UserMixin):
