+++ flask_app/templates/new_article.html
