+++ tests/conftest.py
@@ -1,8 +1,8 @@
 import pytest
 
-from app import create_app
-from app.extensions import db
-from app.models.user import User
+from flask_app import create_app
+from flask_app.extensions import db
+from flask_app.models.user import User
 
 
 @pytest.fixture
