+++ tests/test_post.py
@@ -1,4 +1,4 @@
-from app.models.text import Post, PostReview
+from flask_app.models.text import Post, PostReview
 
 
 def test_posts(client, auth):
